+++ graph.h
@@ -157,7 +157,7 @@ public:
      * @return - if edge exists, return edge wright
      *         - if doesn't, return InvalidWeight
      */
-    int getEdgeWeight(Vertex source, Vertex destination) const;
+    double getEdgeWeight(Vertex source, Vertex destination) const;
 
     /**
      * Inserts a new vertex into the graph and initializes its label as "".
@@ -200,7 +200,7 @@ public:
      * @return - if edge exists, set edge weight and return  edge with new weight
      *         - if not, return InvalidEdge
      */
-    Edge setEdgeWeight(Vertex source, Vertex destination, int weight);
+    Edge setEdgeWeight(Vertex source, Vertex destination, double weight);
 
     void setLatLongPairs(Vertex source, Vertex destination, const pair<double, 
                          double>& sourceCoords, const pair<double,
